dummyData

+++ API/Entities/Product.cs
@@ -4,11 +4,11 @@ namespace API.Entities
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public int Age { get; set; }
+        public double Age { get; set; }
         public string Sex { get; set; }
         public double Milk { get; set; }
         public bool Shaveable { get; set; }
         public string Brand { get; set; }
-        public int DaysLastShaved { get; set; }
+        public double DaysLastShaved { get; set; }
     }
 }
